refactor: rename types to operate_types
